+++ index.html
@@ -189,6 +189,6 @@
         </form>
     </dialog>
     <footer><h4>Made by Adil, Josef & Osman</h4></footer>
-    <script src="js/app.js"></script>
+    <script src="app.js" type="module"></script>
   </body>
 </html>
